+++ actionpack/CHANGELOG.md
@@ -1,7 +1,6 @@
-*   Fix `HostAuthorization` potentially displaying the value of the
-    X_FORWARDED_HOST header when the HTTP_HOST header is being blocked.
+*   Add support for Playwright as a driver for system tests.
 
-    *Hartley McGuire*, *Daniel Schlosser*
+    *Yuki Nishijima*
 
 *   Rename `fixture_file_upload` method to `file_fixture_upload`